Changed primary key type for ArticleEntity and Article

+++ app/src/main/java/com/peeranm/worldnews/feature_news/data/local/dao/NewsDao.kt
@@ -14,7 +14,7 @@ interface NewsDao {
     fun insertAll(articles: List<ArticleEntity>)
 
     @Query("select * from table_articles where id =:id")
-    suspend fun getArticleById(id: Int): ArticleEntity?
+    suspend fun getArticleById(id: Long): ArticleEntity?
 
     @Query("select * from table_articles")
     fun getHeadlinesPagingSource(): PagingSource<Int, ArticleEntity>
